Refactor family and group config cache from `loader.lib` to `tools.lib`
- Plus clarify that the group/family config data is a cache.
